fix: resolve PR169 feedback, enforce zero-warning and LOC gate baseline
- mcp: add #[allow(dead_code)] on McpTools.database (planned-use field,
  MCP handlers are stubs with TODO comments). Follows existing pattern
  from export_payload.rs.
- hyperdim: extract Serialize/Deserialize impls and HVecVisitor to new
  hyperdim_serde.rs (507 → 429).
- framework: extract persist(), persistence_health_check(),
  load_replace(), load_merge() to new framework_persistence.rs
  (510 → 364).
- persistence: extract save_index(), load_index() to new
  persistence_index.rs (532 → 495).
- singularity: extract find_similar(), find_similar_arc(),
  find_similar_cached() to new singularity_search.rs (604 → 462).
  Removed now-unused Arc/Ordering/ScoredCandidateParams imports.
- lib.rs: register hyperdim_serde, framework_persistence,
  persistence_index, and singularity_search modules.

All gates passing: clippy -D warnings, fmt, dead_code check, LOC ≤ 500.
